python: refactor tests for features extraction
